build: adjust system library handling for Windows
Rather than pushing the linked library name in the module map, place it
in the package manifest as the names differ on Windows.  Windows prefers
to statically link libxml2 which requires additional macro definitions
(`LIBXML_STATIC`) and changes the library name to `libxml2s` (`lib` is
used as a prefix for differentiating static libraries and import
libraries).  Since the nuget integration is not yet complete, we require
additional flags to be passed on Windows.  With this change we are able
to build on Windows.